renamed simulation to environment
