+++ pkg/environment/firebase.go
@@ -1,4 +1,4 @@
-package simulation
+package environment
 
 import (
 	"context"
